requirements.txt
academy-py
globus-compute-sdk
emoji
